fix: Minimap is now decent, norm OK
